Trial/Core/Src/requirements.txt
pyserial>=3.5
requests>=2.32.0
solana>=0.36.0
solders>=0.27.0
